% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_flow.R
\name{clip_flow}
\alias{clip_flow}
\title{Clip flow measurements to zero when no rain has fallen}
\usage{
clip_flow(data)
}
\arguments{
\item{data}{A \code{data.frame} containing columns \code{date_time},
\code{flow}, and \code{rain}.}
}
\value{
A \code{data.frame} arranged by \code{date_time} with the flow column
  updated so that flow is 0 when no rain has fallen.
}
\description{
The flumes that measure flow sometimes get stuck. This is clear when there
has been no rainfall recently, but the flume is still recording flow. When
this occurs, we set the flow to zero. This function sorts the
\code{data.frame} by date_time and then calculates the rain in the previous
24*12 rows of the data.frame as this assumes the data are in 5 minute
increments.
}
