% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_flow.R
\name{max_or_na}
\alias{max_or_na}
\title{Returns max or NA (rather than -Inf) if there are no non-NA values}
\usage{
max_or_na(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the result of max or NA if there are no non-NA values
}
\description{
The max function returns -Inf and a warning if there are no non-NA values in
their arguments. This function suppress the warning and returns NA.
}
