% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_sampleID.R
\name{spread_sampleID}
\alias{spread_sampleID}
\title{Assigns a sampleID to the entire rain event.}
\usage{
spread_sampleID(data)
}
\arguments{
\item{A}{\code{data.frame} containing the columns \code{sampleID} and
\code{flow}.}
}
\description{
During a rain event, runoff samples are collected and, later, the sample is
sent to an analytical lab for analysis. These samples need to be attributed
to the entire event
}
\examples{
d <- data.frame(flow = c(0,1,NA,1,0),
                sampleID = c(NA,NA,1,NA,NA))
spread_sampleID(d)

d <- data.frame(flow = c(0,1,NA,1,NA,1,0),
                sampleID = c(NA,NA,1,NA,2,NA,NA))
spread_sampleID(d)

}
